Finished new login
+ New what's knew screen
+ Now only more testing, new load mode and more refactor.
