+++ app/lib/login.dart
@@ -2,8 +2,10 @@ import 'dart:io';
 
 import 'package:flutter/material.dart';
 import 'package:package_info_plus/package_info_plus.dart';
+import 'package:sph_plan/home_page.dart';
 import 'package:sph_plan/shared/apps.dart';
 import 'package:sph_plan/shared/exceptions/client_status_exceptions.dart';
+import 'package:sph_plan/shared/whats_new.dart';
 import 'package:sph_plan/view/bug_report/send_bugreport.dart';
 import 'package:sph_plan/view/login/screen.dart';
 
@@ -85,11 +87,14 @@ class _LoginScreenState extends State<LoginScreen> {
   final List<String> steps = [Step.login];
   late final ProgressNotifier progress;
 
+  ValueNotifier<bool> noInternet = ValueNotifier<bool>(false);
   ValueNotifier<bool> isError = ValueNotifier<bool>(false);
   final Map<String, LanisException?> errors = {Step.login: null};
 
+  ValueNotifier<bool> finishedLoadingStorage = ValueNotifier<bool>(false);
+
   void openWelcomeScreen() {
-    Navigator.push(
+    Navigator.pushReplacement(
       context,
       MaterialPageRoute(builder: (context) => const WelcomeLoginScreen()),
     ).then((_) {
@@ -109,6 +114,11 @@ class _LoginScreenState extends State<LoginScreen> {
         loadingMessage.value = Message.error;
         isError.value = true;
         errors[applet.step] = data.content;
+
+        if (data.content is NoConnectionException) {
+          noInternet.value = true;
+        }
+
         return;
       } else if (data.status == FetcherStatus.done) {
         progress.set(applet.step, Status.finished);
@@ -122,7 +132,6 @@ class _LoginScreenState extends State<LoginScreen> {
 
   Future<void> performLogin() async {
     // Step 1
-    await client.loadFromStorage();
 
     // It doesn't show it immediately but faster than before.
     // I think this check is enough.
@@ -140,14 +149,33 @@ class _LoginScreenState extends State<LoginScreen> {
       client.prepareFetchers();
       progress.set(Step.login, Status.finished);
 
-      // Step 3 (only load vp)
+      // Step 3 (only load vp) will be removed
       if (client.loadMode == "fast") {
         loadingMessage.value = Message.loadFast;
         if (client.doesSupportFeature(SPHAppEnum.vertretungsplan)) {
           await fetchApplet(Applet(fetcher: client.substitutionsFetcher, step: Step.substitutions, finishMessage: "Vertretungen wurden geladen!"));
         }
 
-        loadingMessage.value = Message.finalise;
+        if (!isError.value) {
+          loadingMessage.value = Message.finalise;
+          whatsNew().then((value) {
+            if (value != null) {
+              Navigator.push(
+                context,
+                MaterialPageRoute(builder: (context) => ReleaseNotesScreen(value)),
+              ).then((_) {
+                //client.prepareFetchers();
+                //openFeature(getDefaultFeature());
+              });
+            }
+          });
+
+          // Context is always mounted
+          Navigator.pushReplacement(
+            context,
+            MaterialPageRoute(builder: (context) => const HomePage()),
+          );
+        }
         return;
       }
 
@@ -157,7 +185,26 @@ class _LoginScreenState extends State<LoginScreen> {
       await Future.wait(List.generate(appletFetchers.length, (index) => fetchApplet(appletFetchers[index])));
 
       // Step 5 (finish)
-      loadingMessage.value = Message.finalise;
+      if (!isError.value) {
+        loadingMessage.value = Message.finalise;
+        whatsNew().then((value) {
+          if (value != null) {
+            Navigator.push(
+              context,
+              MaterialPageRoute(builder: (context) => ReleaseNotesScreen(value)),
+            ).then((_) {
+              //client.prepareFetchers();
+              //openFeature(getDefaultFeature());
+            });
+          }
+        });
+
+        // Context is always mounted
+        Navigator.pushReplacement(
+          context,
+          MaterialPageRoute(builder: (context) => const HomePage()),
+        );
+      }
       return;
     } on WrongCredentialsException {
       openWelcomeScreen();
@@ -166,8 +213,13 @@ class _LoginScreenState extends State<LoginScreen> {
     } on LanisException catch (e) {
       errors[Step.login] = e;
       isError.value = true;
-      loadingMessage.value = Message.error;
       progress.set(Step.login, Status.error);
+
+      if (e is NoConnectionException) {
+        noInternet.value = true;
+      }
+
+      loadingMessage.value = Message.error;
     }
   }
 
@@ -199,41 +251,39 @@ class _LoginScreenState extends State<LoginScreen> {
 
   @override
   void initState() {
-    if (client.doesSupportFeature(SPHAppEnum.vertretungsplan)) {
-      steps.add(Step.substitutions);
-      appletFetchers.add(Applet(fetcher: client.substitutionsFetcher, step: Step.substitutions, finishMessage: "Vertretungen wurden geladen!"));
-      errors.addEntries([MapEntry(Step.substitutions, null)]);
-    }
-
-    if (client.loadMode == "full") {
-      if (client.doesSupportFeature(SPHAppEnum.meinUnterricht)) {
-        steps.add(Step.meinUnterricht);
-        appletFetchers.add(Applet(fetcher: client.meinUnterrichtFetcher, step: Step.meinUnterricht, finishMessage: "Mein Unterricht wurde geladen!"));
-        errors.addEntries([MapEntry(Step.meinUnterricht, null)]);
-      }
-      if (client.doesSupportFeature(SPHAppEnum.nachrichten)) {
-        steps.add(Step.conversations);
-        appletFetchers.add(Applet(fetcher: client.invisibleConversationsFetcher, step: Step.conversations, finishMessage: "Nachrichten wurden geladen! (1/2)"));
-        appletFetchers.add(Applet(fetcher: client.visibleConversationsFetcher, step: Step.conversations, finishMessage: "Nachrichten wurden geladen! (2/2)"));
-        errors.addEntries([MapEntry(Step.conversations, null)]);
-      }
-      if (client.doesSupportFeature(SPHAppEnum.kalender)) {
-        steps.add(Step.calendar);
-        appletFetchers.add(Applet(fetcher: client.calendarFetcher, step: Step.calendar, finishMessage: "Kalender wurde geladen!"));
-        errors.addEntries([MapEntry(Step.calendar, null)]);
-      }
-    }
-
-    progress = ProgressNotifier(steps);
-
     WidgetsBinding.instance.addPostFrameCallback((_) {
-      performLogin();
+      client.loadFromStorage().then((_) {
+        finishedLoadingStorage.value = true;
 
-      /*whatsNew().then((value) {
-        if (value != null) {
-          openReleaseNotesModal(context, value);
+        if (client.doesSupportFeature(SPHAppEnum.vertretungsplan)) {
+          steps.add(Step.substitutions);
+          appletFetchers.add(Applet(fetcher: client.substitutionsFetcher, step: Step.substitutions, finishMessage: "Vertretungen wurden geladen!"));
+          errors.addEntries([MapEntry(Step.substitutions, null)]);
         }
-      });*/
+
+        if (client.loadMode == "full") {
+          if (client.doesSupportFeature(SPHAppEnum.meinUnterricht)) {
+            steps.add(Step.meinUnterricht);
+            appletFetchers.add(Applet(fetcher: client.meinUnterrichtFetcher, step: Step.meinUnterricht, finishMessage: "Mein Unterricht wurde geladen!"));
+            errors.addEntries([MapEntry(Step.meinUnterricht, null)]);
+          }
+          if (client.doesSupportFeature(SPHAppEnum.nachrichten)) {
+            steps.add(Step.conversations);
+            appletFetchers.add(Applet(fetcher: client.invisibleConversationsFetcher, step: Step.conversations, finishMessage: "Nachrichten wurden geladen! (1/2)"));
+            appletFetchers.add(Applet(fetcher: client.visibleConversationsFetcher, step: Step.conversations, finishMessage: "Nachrichten wurden geladen! (2/2)"));
+            errors.addEntries([MapEntry(Step.conversations, null)]);
+          }
+          if (client.doesSupportFeature(SPHAppEnum.kalender)) {
+            steps.add(Step.calendar);
+            appletFetchers.add(Applet(fetcher: client.calendarFetcher, step: Step.calendar, finishMessage: "Kalender wurde geladen!"));
+            errors.addEntries([MapEntry(Step.calendar, null)]);
+          }
+        }
+
+        progress = ProgressNotifier(steps);
+
+        performLogin();
+      });
     });
 
     super.initState();
@@ -274,10 +324,21 @@ class _LoginScreenState extends State<LoginScreen> {
                       borderRadius: BorderRadius.circular(16.0)
                   ),
                   padding: const EdgeInsets.all(12.0),
-                  child: Text(
-                      "Willkommen zurück!",
-                      style: Theme.of(context).textTheme.headlineMedium?.copyWith(color: Theme.of(context).colorScheme.onPrimary)
-                  ),
+                  child: ValueListenableBuilder(
+                    valueListenable: noInternet,
+                    builder: (context, _noInternet, _) {
+                      if (_noInternet) {
+                        return Text(
+                            "Kein Internet!",
+                            style: Theme.of(context).textTheme.headlineMedium?.copyWith(color: Theme.of(context).colorScheme.onPrimary)
+                        );
+                      }
+                      return Text(
+                          "Willkommen zurück!",
+                          style: Theme.of(context).textTheme.headlineMedium?.copyWith(color: Theme.of(context).colorScheme.onPrimary)
+                      );
+                    },
+                  )
                 ),
                 Padding(
                   padding: const EdgeInsets.only(top: 8.0),
@@ -291,84 +352,108 @@ class _LoginScreenState extends State<LoginScreen> {
                       child: Column(
                         crossAxisAlignment: CrossAxisAlignment.start,
                         children: [
-                          ListenableBuilder(
-                              listenable: progress,
-                              builder: (context, _) {
-                                return ListView.separated(
-                                  shrinkWrap: true,
-                                    itemCount: steps.length,
-                                    separatorBuilder: (context, index) {
-                                      return const SizedBox(
-                                        height: 8,
-                                      );
-                                    },
-                                    itemBuilder: (context, index) {
-                                      final String step = steps[index];
-                                      final Status status = progress.steps[step]!;
-                                      return Row(
-                                        mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                                        children: [
-                                          Row(
-                                            mainAxisSize: MainAxisSize.min,
-                                            children: [
-                                              Padding(
-                                                padding: const EdgeInsets.only(right: 4.0),
-                                                child: Icon(
-                                                  getIcon(status),
-                                                  color: Theme.of(context).colorScheme.onSecondary,
-                                                ),
-                                              ),
-                                              Text(
-                                                  step,
-                                                  style: Theme.of(context).textTheme.labelLarge?.copyWith(color: Theme.of(context).colorScheme.onSecondary)
-                                              ),
-                                            ],
-                                          ),
-                                          Text(
-                                              getProgressMessage(status),
-                                              style: Theme.of(context).textTheme.bodyMedium?.copyWith(color: Theme.of(context).colorScheme.onSecondary)
-                                          )
-                                        ],
-                                      );
-                                    }
+                          ValueListenableBuilder(
+                              valueListenable: finishedLoadingStorage,
+                              builder: (context, _finishedLoadingStorage, _) {
+                                if (_finishedLoadingStorage) {
+                                  return ListenableBuilder(
+                                      listenable: progress,
+                                      builder: (context, _) {
+                                        return ListView.separated(
+                                            shrinkWrap: true,
+                                            itemCount: steps.length,
+                                            separatorBuilder: (context, index) {
+                                              return const SizedBox(
+                                                height: 8,
+                                              );
+                                            },
+                                            itemBuilder: (context, index) {
+                                              final String step = steps[index];
+                                              final Status status = progress.steps[step]!;
+                                              return Row(
+                                                mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                                                children: [
+                                                  Row(
+                                                    mainAxisSize: MainAxisSize.min,
+                                                    children: [
+                                                      Padding(
+                                                        padding: const EdgeInsets.only(right: 4.0),
+                                                        child: Icon(
+                                                          getIcon(status),
+                                                          color: Theme.of(context).colorScheme.onSecondary,
+                                                        ),
+                                                      ),
+                                                      Text(
+                                                          step,
+                                                          style: Theme.of(context).textTheme.labelLarge?.copyWith(color: Theme.of(context).colorScheme.onSecondary)
+                                                      ),
+                                                    ],
+                                                  ),
+                                                  Text(
+                                                      getProgressMessage(status),
+                                                      style: Theme.of(context).textTheme.bodyMedium?.copyWith(color: Theme.of(context).colorScheme.onSecondary)
+                                                  )
+                                                ],
+                                              );
+                                            }
+                                        );
+                                      }
+                                  );
+                                }
+                                return Row(
+                                  mainAxisAlignment: MainAxisAlignment.center,
+                                  children: [
+                                    Text(
+                                        "Lade gespeicherte Daten...",
+                                        style: Theme.of(context).textTheme.bodyMedium?.copyWith(color: Theme.of(context).colorScheme.onSecondary)
+                                    )
+                                  ],
                                 );
                               }
-                          ),
+                          )
                         ],
                       )
                   ),
                 ),
                 ValueListenableBuilder(
                     valueListenable: isError,
-                    builder: (context, value, _) {
-                      if (value) {
+                    builder: (context, _isError, _) {
+                      if (_isError) {
                         return Padding(
                           padding: const EdgeInsets.only(top: 16.0),
                           child: Row(
                             mainAxisAlignment: MainAxisAlignment.center,
                             children: [
-                              FilledButton(
-                                  onPressed: () {
-                                    // Prepare error message
-                                    String errorInfo = "";
-                                    errors.forEach((key, value) {
-                                      if (value == null) {
-                                        return;
-                                      }
-                                      errorInfo += "$key - ${value.cause}\n";
-                                    });
+                              ValueListenableBuilder(
+                                  valueListenable: noInternet,
+                                  builder: (context, _noInternet, _) {
+                                    if (_noInternet) {
+                                      return const SizedBox.shrink();
+                                    }
+                                    return FilledButton(
+                                        onPressed: () {
+                                          // Prepare error message
+                                          String errorInfo = "";
+                                          errors.forEach((key, value) {
+                                            if (value == null) {
+                                              return;
+                                            }
+                                            errorInfo += "$key - ${value.cause}\n";
+                                          });
 
-                                    Navigator.push(
-                                      context,
-                                      MaterialPageRoute(
-                                          builder: (context) =>
-                                              BugReportScreen(
-                                                  generatedMessage:
-                                                  "AUTOMATISCH GENERIERT (LOGIN PAGE):\n$errorInfo\nMehr Details von dir:\n")),
-                                    );
-                                  },
-                                  child: const Text(
-                                      "Fehlerbericht senden")),
+                                          Navigator.push(
+                                            context,
+                                            MaterialPageRoute(
+                                                builder: (context) =>
+                                                    BugReportScreen(
+                                                        generatedMessage:
+                                                        "AUTOMATISCH GENERIERT (LOGIN PAGE):\n$errorInfo\nMehr Details von dir:\n")),
+                                          );
+                                        },
+                                        child: const Text(
+                                            "Fehlerbericht senden"));
+                                  }
+                              ),
                               Padding(
                                 padding: const EdgeInsets.only(left: 8),
                                 child: OutlinedButton(
@@ -377,6 +462,7 @@ class _LoginScreenState extends State<LoginScreen> {
                                       progress.reset();
                                       errors.updateAll((key, value) => value = null);
                                       isError.value = false;
+                                      noInternet.value = false;
                                       loadingMessage.value = "Initialisieren...";
 
                                       await performLogin();
@@ -408,11 +494,11 @@ class _LoginScreenState extends State<LoginScreen> {
                     children: [
                       ValueListenableBuilder(
                         valueListenable: loadingMessage,
-                        builder: (context, value, _) {
+                        builder: (context, _loadingMessage, _) {
                           return Padding(
                             padding: const EdgeInsets.only(right: 12.0),
                             child: Text(
-                                value,
+                                _loadingMessage,
                                 style: Theme.of(context).textTheme.labelLarge?.copyWith(color: Theme.of(context).colorScheme.onPrimary)
                             ),
                           );
@@ -420,8 +506,8 @@ class _LoginScreenState extends State<LoginScreen> {
                       ),
                       ValueListenableBuilder(
                           valueListenable: isError,
-                          builder: (context, value, _) {
-                            if (value) {
+                          builder: (context, _isError, _) {
+                            if (_isError) {
                               return Icon(
                                 Icons.error,
                                 color: Theme.of(context).colorScheme.onPrimary,
